+++ packages/edit-site/src/components/global-styles/ui.js
@@ -1,0 +1,79 @@
+/**
+ * WordPress dependencies
+ */
+import {
+	__experimentalNavigatorProvider as NavigatorProvider,
+	__experimentalNavigatorScreen as NavigatorScreen,
+} from '@wordpress/components';
+import { getBlockTypes } from '@wordpress/blocks';
+
+/**
+ * Internal dependencies
+ */
+import ScreenRoot from './screen-root';
+import ScreenBlockList from './screen-block-list';
+import ScreenBlock from './screen-block';
+import ScreenTypography from './screen-typography';
+import ScreenColors from './screen-colors';
+import ScreenColorPalette from './screen-color-palette';
+import ScreenLayout from './screen-layout';
+
+function ContextScreens( { name } ) {
+	const parentMenu = name === undefined ? '' : '/blocks/' + name;
+
+	return (
+		<>
+			<NavigatorScreen path={ parentMenu + '/typography' }>
+				<ScreenTypography name={ name } />
+			</NavigatorScreen>
+
+			<NavigatorScreen path={ parentMenu + '/colors' }>
+				<ScreenColors name={ name } />
+			</NavigatorScreen>
+
+			<NavigatorScreen path={ parentMenu + '/colors/palette' }>
+				<ScreenColorPalette name={ name } />
+			</NavigatorScreen>
+
+			<NavigatorScreen path={ parentMenu + '/layout' }>
+				<ScreenLayout name={ name } />
+			</NavigatorScreen>
+		</>
+	);
+}
+
+function GlobalStylesUI() {
+	const blocks = getBlockTypes();
+
+	return (
+		<NavigatorProvider initialPath="/">
+			<NavigatorScreen path="/">
+				<ScreenRoot />
+			</NavigatorScreen>
+
+			<NavigatorScreen path="/blocks">
+				<ScreenBlockList />
+			</NavigatorScreen>
+
+			{ blocks.map( ( block ) => (
+				<NavigatorScreen
+					key={ 'menu-block-' + block.name }
+					path={ '/blocks/' + block.name }
+				>
+					<ScreenBlock name={ block.name } />
+				</NavigatorScreen>
+			) ) }
+
+			<ContextScreens />
+
+			{ blocks.map( ( block ) => (
+				<ContextScreens
+					key={ 'screens-block-' + block.name }
+					name={ block.name }
+				/>
+			) ) }
+		</NavigatorProvider>
+	);
+}
+
+export default GlobalStylesUI;
